replace govalidator with go-playground/validator
