+++ app/formdata/form_data.go
@@ -5,12 +5,17 @@ import (
 	"net/http"
 	"reflect"
 	"regexp"
+	"strconv"
 	"strings"
 	"time"
 	"unicode"
 	"unicode/utf8"
 
-	"github.com/France-ioi/govalidator"
+	english "github.com/go-playground/locales/en"
+	ut "github.com/go-playground/universal-translator"
+	"gopkg.in/go-playground/validator.v9"
+	en "gopkg.in/go-playground/validator.v9/translations/en"
+
 	"github.com/France-ioi/mapstructure"
 )
 
@@ -20,13 +25,54 @@ type FormData struct {
 	fieldErrors         FieldErrors
 	metadata            mapstructure.Metadata
 	usedKeys            map[string]bool
+
+	validate *validator.Validate
+	trans    ut.Translator
 }
 
 // NewFormData creates a new FormData object for given definitions
 func NewFormData(definitionStructure interface{}) *FormData {
-	return &FormData{
+	validate := validator.New()
+	var eng = english.New()
+	var uni = ut.New(eng, eng)
+	trans, _ := uni.GetTranslator("en")
+	_ = en.RegisterDefaultTranslations(validate, trans)
+	validate.RegisterTagNameFunc(func(fld reflect.StructField) string {
+		name := strings.SplitN(fld.Tag.Get("json"), ",", 2)[0]
+		if name == "-" {
+			return ""
+		}
+		return name
+	})
+
+	formData := &FormData{
 		definitionStructure: definitionStructure,
+		validate:            validate,
+		trans:               trans,
 	}
+	formData.RegisterValidation("duration", validator.Func(validateDuration))
+	formData.RegisterTranslation("duration", "invalid duration")
+
+	formData.RegisterValidation("dmy-date", validator.Func(validateDMYDate))
+	formData.RegisterTranslation("dmy-date", "should be dd-mm-yyyy")
+
+	return formData
+}
+
+// RegisterValidation adds a validation with the given tag
+func (f *FormData) RegisterValidation(tag string, fn validator.Func) {
+	_ = f.validate.RegisterValidation(tag, fn)
+}
+
+// RegisterTranslation registers translations against the provided tag
+func (f *FormData) RegisterTranslation(tag, text string) {
+	_ = f.validate.RegisterTranslation(tag, f.trans,
+		func(ut ut.Translator) (err error) {
+			return ut.Add(tag, text, false)
+		}, func(ut ut.Translator, fe validator.FieldError) string {
+			t, _ := ut.T(fe.Tag(), fe.Field())
+			return t
+		})
 }
 
 // ParseJSONRequestData parses and validates JSON from the request according to the structure definition
@@ -62,6 +108,29 @@ func (f *FormData) ConstructMapForDB() map[string]interface{} {
 	return result
 }
 
+// ConstructPartialMapForDB constructs a map for updating DB. It uses both the definition and the JSON input
+func (f *FormData) ConstructPartialMapForDB(part string) map[string]interface{} {
+	result := map[string]interface{}{}
+	partField := reflect.ValueOf(f.definitionStructure).Elem().FieldByName(part)
+	partFieldType, _ := reflect.TypeOf(f.definitionStructure).Elem().FieldByName(part)
+	prefix := getJSONFieldName(&partFieldType)
+	if partField.Kind() == reflect.Ptr {
+		partField = partField.Elem()
+	}
+	f.addDBFieldsIntoMap(result, partField, prefix)
+	return result
+}
+
+// IsSet returns true if the field is set
+func (f *FormData) IsSet(key string) bool {
+	return f.usedKeys[key]
+}
+
+// IsValid returns true if the field is valid (there are no errors for this field)
+func (f *FormData) IsValid(key string) bool {
+	return len(f.fieldErrors[key]) == 0
+}
+
 var mapstructTypeErrorRegexp = regexp.MustCompile(`^'([^']*)'\s+(.*)$`)
 var mapstructDecodingErrorRegexp = regexp.MustCompile(`^error decoding '([^']*)':\s+(.*)$`)
 
@@ -85,6 +154,7 @@ func (f *FormData) decodeMapIntoStruct(m map[string]interface{}) {
 		DecodeHook: mapstructure.ComposeDecodeHookFunc(
 			mapstructure.StringToTimeHookFunc(time.RFC3339),
 			toAnythingHookFunc(),
+			stringToInt64HookFunc(),
 		),
 		ErrorUnused:      false, // we will check this on our own
 		Metadata:         &f.metadata,
@@ -106,10 +176,12 @@ func (f *FormData) decodeMapIntoStruct(m map[string]interface{}) {
 				value := make([]byte, len(matches[2]))
 				copy(value, matches[2])
 				f.fieldErrors[string(key)] = append(f.fieldErrors[string(key)], string(value))
+				f.usedKeys[string(key)] = true
 			} else if matches := mapstructDecodingErrorRegexp.FindStringSubmatch(fieldErrorString); len(matches) > 0 {
 				key := make([]byte, len(matches[1]))
 				copy(key, matches[1])
 				f.fieldErrors[string(key)] = append(f.fieldErrors[string(key)], "decoding error: "+matches[2])
+				f.usedKeys[string(key)] = true
 			} else {
 				f.fieldErrors[""] = append(f.fieldErrors[""], fieldErrorString) // should never happen
 			}
@@ -118,38 +190,30 @@ func (f *FormData) decodeMapIntoStruct(m map[string]interface{}) {
 }
 
 func (f *FormData) validateFieldValues() {
-	if _, err := govalidator.ValidateStruct(f.definitionStructure); err != nil {
-		f.processGovalidatorErrors(err)
+	if err := f.validate.Struct(f.definitionStructure); err != nil {
+		f.processValidatorErrors(err)
 	}
 }
 
-func (f *FormData) processGovalidatorErrors(err error) {
-	validatorErrors := err.(govalidator.Errors)
-	for _, validatorError := range validatorErrors {
-		if err, ok := validatorError.(govalidator.Error); ok {
-			currentFieldType := reflect.TypeOf(f.definitionStructure)
-			for pathIndex, pathElement := range err.Path {
-				for currentFieldType.Kind() == reflect.Ptr {
-					currentFieldType = currentFieldType.Elem()
-				}
-				field, _ := currentFieldType.FieldByName(pathElement)
-				currentFieldType = field.Type
-				jsonName := getJSONFieldName(&field)
-				if jsonName != "-" {
-					err.Path[pathIndex] = jsonName
-				}
-			}
+const required = "required"
 
-			path := strings.Join(err.Path, ".")
-			if len(path) > 0 {
-				path += "."
-			}
-			path += err.Name
-			if f.usedKeys[path] {
-				f.fieldErrors[path] = append(f.fieldErrors[path], err.Err.Error())
-			}
-		} else {
-			f.processGovalidatorErrors(validatorError)
+func (f *FormData) processValidatorErrors(err error) {
+	validatorErrors := err.(validator.ValidationErrors)
+	for _, validatorError := range validatorErrors {
+		path := validatorError.Namespace()
+		prefix := ""
+		structType := reflect.TypeOf(f.definitionStructure)
+		for structType.Kind() == reflect.Ptr {
+			structType = structType.Elem()
+		}
+		structName := structType.Name()
+		if structName != "" {
+			prefix = structName + "."
+		}
+		path = strings.TrimPrefix(path, prefix)
+		if f.usedKeys[path] && validatorError.Tag() != required {
+			errorMsg := validatorError.Translate(f.trans)
+			f.fieldErrors[path] = append(f.fieldErrors[path], errorMsg)
 		}
 	}
 }
@@ -161,6 +225,20 @@ func (f *FormData) checkProvidedFields() {
 	for _, usedKey := range f.metadata.Keys {
 		f.usedKeys[usedKey] = true
 	}
+
+	traverseStructure(func(field reflect.Value, structField reflect.StructField, jsonName string) bool {
+		validTag := structField.Tag.Get("validate")
+		tags := strings.Split(validTag, ",")
+		for _, value := range tags {
+			if value == required {
+				if !f.usedKeys[jsonName] {
+					f.fieldErrors[jsonName] = append(f.fieldErrors[jsonName], "missing field")
+				}
+				break
+			}
+		}
+		return true
+	}, reflect.ValueOf(f.definitionStructure), "")
 }
 
 func (f *FormData) addDBFieldsIntoMap(resultMap map[string]interface{}, reflValue reflect.Value, prefix string) {
@@ -256,5 +334,19 @@ func toAnythingHookFunc() mapstructure.DecodeHookFunc {
 		}
 		bytes, _ := json.Marshal(data)
 		return *AnythingFromBytes(bytes), nil
+	}
+}
+
+// stringToInt64HookFunc returns a DecodeHookFunc that converts
+// any value to payloads.Anything.
+func stringToInt64HookFunc() mapstructure.DecodeHookFunc {
+	return func(
+		f reflect.Type,
+		t reflect.Type,
+		data interface{}) (interface{}, error) {
+		if t.Kind() != reflect.Int64 || f.Kind() != reflect.String {
+			return data, nil
+		}
+		return strconv.ParseInt(data.(string), 10, 64)
 	}
 }
