+++ app/formdata/form_data_export_test.go
@@ -5,3 +5,7 @@ import "github.com/France-ioi/mapstructure"
 func ToAnythingHookFunc() mapstructure.DecodeHookFunc {
 	return toAnythingHookFunc()
 }
+
+func StringToInt64HookFunc() mapstructure.DecodeHookFunc {
+	return stringToInt64HookFunc()
+}
