+++ app/payloads/hint_token.go
@@ -11,7 +11,7 @@ import (
 
 // HintToken represents data inside a hint token
 type HintToken struct {
-	Date        string            `json:"date" valid:"matches(^[0-3][0-9]-[0-1][0-9]-\\d{4}$)"` // dd-mm-yyyy
+	Date        string            `json:"date" validate:"dmy-date"` // dd-mm-yyyy
 	UserID      string            `json:"idUser,omitempty"`
 	ItemID      *string           `json:"idItem,omitempty"`
 	LocalItemID string            `json:"idItemLocal"`
