+++ app/payloads/score_token.go
@@ -9,7 +9,7 @@ import (
 // ScoreToken represents data inside a score token
 type ScoreToken struct {
 	// Nullable fields are of pointer types
-	Date         string  `json:"date" valid:"matches(^[0-3][0-9]-[0-1][0-9]-\\d{4}$)"` // dd-mm-yyyy
+	Date         string  `json:"date" validate:"dmy-date"` // dd-mm-yyyy
 	UserID       string  `json:"idUser"`
 	ItemID       *string `json:"idItem,omitempty"`
 	LocalItemID  string  `json:"idItemLocal"`
