+++ go.mod
@@ -6,16 +6,18 @@ require (
 	github.com/CloudyKit/jet v2.1.3-0.20180809161101-62edd43e4f88+incompatible
 	github.com/DATA-DOG/go-sqlmock v1.3.2
 	github.com/DATA-DOG/godog v0.7.13
-	github.com/France-ioi/govalidator v0.0.0-20190304102511-a2eeeae8d4ba
 	github.com/France-ioi/mapstructure v1.1.3-0.20190228185851-70b68b9b4003
 	github.com/akrylysov/algnhsa v0.0.0-20190203201208-70f315bb89bd
 	github.com/aws/aws-lambda-go v1.8.2
 	github.com/go-chi/chi v3.3.3+incompatible
 	github.com/go-chi/cors v1.0.0
 	github.com/go-chi/render v1.0.1
+	github.com/go-playground/locales v0.12.1
+	github.com/go-playground/universal-translator v0.16.0
 	github.com/go-sql-driver/mysql v1.4.1
 	github.com/jinzhu/gorm v1.9.2
 	github.com/jinzhu/inflection v0.0.0-20180308033659-04140366298a // indirect
+	github.com/leodido/go-urn v1.1.0 // indirect
 	github.com/lithammer/dedent v1.1.0
 	github.com/luna-duclos/instrumentedsql v0.0.0-20181127104832-b7d587d28109
 	github.com/pkg/errors v0.8.1 // indirect
@@ -25,6 +27,7 @@ require (
 	github.com/spf13/cobra v0.0.3
 	github.com/spf13/viper v1.3.1
 	github.com/stretchr/testify v1.2.2
+	gopkg.in/go-playground/validator.v9 v9.29.0
 	gopkg.in/gorp.v1 v1.7.2 // indirect
 	gopkg.in/jose.v1 v1.0.0-20161127122323-a941c3995164
 	gopkg.in/yaml.v2 v2.2.2
